popcnt -> bits.OnesCount64
