+++ popcnt.go
@@ -1,20 +1,13 @@
 package roaring
 
-// bit population count, take from
-// https://code.google.com/p/go/issues/detail?id=4988#c11
-// credit: https://code.google.com/u/arnehormann/
-// credit: https://play.golang.org/p/U7SogJ7psJ
-// credit: http://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetParallel
+import "math/bits"
+
+// bit population count, uses Go 1.9 math/bits library
 func popcount(x uint64) uint64 {
-	x -= (x >> 1) & 0x5555555555555555
-	x = (x>>2)&0x3333333333333333 + x&0x3333333333333333
-	x += x >> 4
-	x &= 0x0f0f0f0f0f0f0f0f
-	x *= 0x0101010101010101
-	return x >> 56
+	return uint64(bits.OnesCount64(x))
 }
 
-func popcntSliceGo(s []uint64) uint64 {
+func popcntSlice(s []uint64) uint64 {
 	cnt := uint64(0)
 	for _, x := range s {
 		cnt += popcount(x)
@@ -22,7 +15,7 @@ func popcntSliceGo(s []uint64) uint64 {
 	return cnt
 }
 
-func popcntMaskSliceGo(s, m []uint64) uint64 {
+func popcntMaskSlice(s, m []uint64) uint64 {
 	cnt := uint64(0)
 	for i := range s {
 		cnt += popcount(s[i] &^ m[i])
@@ -30,7 +23,7 @@ func popcntMaskSliceGo(s, m []uint64) uint64 {
 	return cnt
 }
 
-func popcntAndSliceGo(s, m []uint64) uint64 {
+func popcntAndSlice(s, m []uint64) uint64 {
 	cnt := uint64(0)
 	for i := range s {
 		cnt += popcount(s[i] & m[i])
@@ -38,7 +31,7 @@ func popcntAndSliceGo(s, m []uint64) uint64 {
 	return cnt
 }
 
-func popcntOrSliceGo(s, m []uint64) uint64 {
+func popcntOrSlice(s, m []uint64) uint64 {
 	cnt := uint64(0)
 	for i := range s {
 		cnt += popcount(s[i] | m[i])
@@ -46,7 +39,7 @@ func popcntOrSliceGo(s, m []uint64) uint64 {
 	return cnt
 }
 
-func popcntXorSliceGo(s, m []uint64) uint64 {
+func popcntXorSlice(s, m []uint64) uint64 {
 	cnt := uint64(0)
 	for i := range s {
 		cnt += popcount(s[i] ^ m[i])
