+++ popcnt_test.go
@@ -10,69 +10,49 @@ import (
 
 func TestPopcntSlice(t *testing.T) {
 	s := []uint64{2, 3, 5, 7, 11, 13, 17, 19, 23, 29}
-	resGo := popcntSliceGo(s)
-	resAsm := popcntSliceAsm(s)
-	if resGo != resAsm {
-		t.Errorf("The implementations are different: GO %d != ASM %d", resGo, resAsm)
-	}
 	res := popcntSlice(s)
-	if res != resGo {
-		t.Errorf("The implementations are different")
+	expected := uint64(27)
+	if res != expected {
+		t.Errorf("Got %d, expected %d", res, expected)
 	}
 }
 
 func TestPopcntMaskSlice(t *testing.T) {
 	s := []uint64{2, 3, 5, 7, 11, 13, 17, 19, 23, 29}
 	m := []uint64{31, 37, 41, 43, 47, 53, 59, 61, 67, 71}
-	resGo := popcntMaskSliceGo(s, m)
-	resAsm := popcntMaskSliceAsm(s, m)
-	if resGo != resAsm {
-		t.Errorf("The implementations are different: GO %d != ASM %d", resGo, resAsm)
-	}
 	res := popcntMaskSlice(s, m)
-	if res != resGo {
-		t.Errorf("The implementations are different")
+	expected := uint64(9)
+	if res != expected {
+		t.Errorf("Got %d, expected %d", res, expected)
 	}
 }
 
 func TestPopcntAndSlice(t *testing.T) {
 	s := []uint64{2, 3, 5, 7, 11, 13, 17, 19, 23, 29}
 	m := []uint64{31, 37, 41, 43, 47, 53, 59, 61, 67, 71}
-	resGo := popcntAndSliceGo(s, m)
-	resAsm := popcntAndSliceAsm(s, m)
-	if resGo != resAsm {
-		t.Errorf("The implementations are different: GO %d != ASM %d", resGo, resAsm)
-	}
 	res := popcntAndSlice(s, m)
-	if res != resGo {
-		t.Errorf("The implementations are different")
+	expected := uint64(18)
+	if res != expected {
+		t.Errorf("Got %d, expected %d", res, expected)
 	}
 }
 
 func TestPopcntOrSlice(t *testing.T) {
 	s := []uint64{2, 3, 5, 7, 11, 13, 17, 19, 23, 29}
 	m := []uint64{31, 37, 41, 43, 47, 53, 59, 61, 67, 71}
-	resGo := popcntOrSliceGo(s, m)
-	resAsm := popcntOrSliceAsm(s, m)
-	if resGo != resAsm {
-		t.Errorf("The implementations are different: GO %d != ASM %d", resGo, resAsm)
-	}
 	res := popcntOrSlice(s, m)
-	if res != resGo {
-		t.Errorf("The implementations are different")
+	expected := uint64(50)
+	if res != expected {
+		t.Errorf("Got %d, expected %d", res, expected)
 	}
 }
 
 func TestPopcntXorSlice(t *testing.T) {
 	s := []uint64{2, 3, 5, 7, 11, 13, 17, 19, 23, 29}
 	m := []uint64{31, 37, 41, 43, 47, 53, 59, 61, 67, 71}
-	resGo := popcntXorSliceGo(s, m)
-	resAsm := popcntXorSliceAsm(s, m)
-	if resGo != resAsm {
-		t.Errorf("The implementations are different: GO %d != ASM %d", resGo, resAsm)
-	}
 	res := popcntXorSlice(s, m)
-	if res != resGo {
-		t.Errorf("The implementations are different")
+	expected := uint64(32)
+	if res != expected {
+		t.Errorf("Got %d, expected %d", res, expected)
 	}
 }
